refactor: move ddtrace.profiling._attr to ddtrace.utils (#2230)
