+++ ddtrace/profiling/exporter/http.py
@@ -12,9 +12,9 @@ import ddtrace
 from ddtrace.internal import agent
 from ddtrace.internal import runtime
 from ddtrace.internal.runtime import container
-from ddtrace.profiling import _attr
 from ddtrace.profiling import exporter
 from ddtrace.profiling.exporter import pprof
+from ddtrace.utils import attr as attr_utils
 from ddtrace.utils.formats import parse_tags_str
 from ddtrace.vendor import attr
 from ddtrace.vendor import six
@@ -39,7 +39,7 @@ class PprofHTTPExporter(pprof.PprofExporter):
 
     endpoint = attr.ib()
     api_key = attr.ib(default=None)
-    timeout = attr.ib(factory=_attr.from_env("DD_PROFILING_API_TIMEOUT", agent.DEFAULT_TIMEOUT, float), type=float)
+    timeout = attr.ib(factory=attr_utils.from_env("DD_PROFILING_API_TIMEOUT", agent.DEFAULT_TIMEOUT, float), type=float)
     service = attr.ib(default=None)
     env = attr.ib(default=None)
     version = attr.ib(default=None)
