+++ ddtrace/utils/attr.py
